relpath
This patch adds an utility for computing relative path according to a given
base directory. This is useful for libraries that are to install .lpk in
order to be used with lazarus. As these libraries may be compiled without
need to build depend on Lazarus, we add it in fp-utils package which is likely
to be used by any big project which build depends on fp-compiler.


Gbp-Pq: Name relpath.patch